requirements.txt
django-extensions
geopy
